+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Upcoming
+- added `pattern.js` tests
+
 
 ## 0.0.1
 - first release